requirements.txt
wheel
flask
pymongo
pyhumps
simple-websocket
git+https://github.com/seraphli/od-python.git
flask-socketio
tqdm
requests
simplejson
numpy
matplotlib
pandas
Pillow
pytesseract